+++ composer.json
@@ -13,9 +13,10 @@
     "require": {
         "php": "^8.1",
         "ui-awesome/html-attribute": "^0.6@dev",
+        "ui-awesome/html-contracts": "^0.1@dev",
         "ui-awesome/html-helper": "^0.7",
-        "ui-awesome/html-interop": "^0.3",
-        "ui-awesome/html-mixin": "^0.4"
+        "ui-awesome/html-interop": "^0.4@dev",
+        "ui-awesome/html-mixin": "dev-main as 0.4"
     },
     "require-dev": {
         "infection/infection": "^0.27|^0.32",
