+++ src/Element/BaseBlock.php
@@ -4,6 +4,11 @@ declare(strict_types=1);
 
 namespace UIAwesome\Html\Core\Element;
 
+use BackedEnum;
+use Fiber;
+use LogicException;
+use RuntimeException;
+use stdClass;
 use UIAwesome\Html\Attribute\Global\{
     CanBeAutofocus,
     CanBeHidden,
@@ -25,23 +30,28 @@ use UIAwesome\Html\Attribute\Global\{
     HasTitle,
     HasTranslate,
 };
+use UIAwesome\Html\Contracts\Attribute\AttributesInterface;
+use UIAwesome\Html\Contracts\Element\BlockInterface;
 use UIAwesome\Html\Core\Base\BaseTag;
+use UIAwesome\Html\Core\Exception\Message;
 use UIAwesome\Html\Core\Html;
 use UIAwesome\Html\Helper\LineBreakNormalizer;
-use UIAwesome\Html\Interop\BlockInterface;
 use UIAwesome\Html\Mixin\{HasAttributes, HasContent};
+use WeakMap;
+
+use function array_pop;
 
 /**
  * Provides the base implementation for block HTML elements.
  *
- * Subclasses return a {@see BlockInterface} tag and inherit attribute and content handling.
+ * Subclasses return a {@see BackedEnum} tag and inherit attribute and content handling.
  *
  * @link https://developer.mozilla.org/en-US/docs/Glossary/Block-level_content
  *
  * @copyright Copyright (C) 2025 Terabytesoftw.
  * @license https://opensource.org/license/bsd-3-clause BSD 3-Clause License.
  */
-abstract class BaseBlock extends BaseTag
+abstract class BaseBlock extends BaseTag implements AttributesInterface, BlockInterface
 {
     use CanBeAutofocus;
     use CanBeHidden;
@@ -66,19 +76,93 @@ abstract class BaseBlock extends BaseTag
     use HasTranslate;
 
     /**
+     * Tracks whether `begin()` was executed for this instance.
+     */
+    private bool $beginExecuted = false;
+
+    /**
+     * Stores the sentinel object for the main execution context.
+     */
+    private static stdClass|null $mainThread = null;
+
+    /**
+     * Stores per-context begin/end stacks.
+     *
+     * @phpstan-var WeakMap<object, static[]>|null
+     */
+    private static WeakMap|null $stack = null;
+
+    /**
      * Returns the tag instance representing the block element.
      *
      * Usage example:
      * ```php
-     * public function getTag(): BlockInterface
+     * public function getTag(): BackedEnum
      * {
      *     return \UIAwesome\Html\Interop\Block::DIV;
      * }
      * ```
      *
-     * @return BlockInterface Tag instance for the block element.
+     * @return BackedEnum Tag instance for the block element.
      */
-    abstract protected function getTag(): BlockInterface;
+    abstract protected function getTag(): BackedEnum;
+
+    /**
+     * Starts begin/end rendering for this instance.
+     *
+     * @return string Empty string for output buffering compatibility.
+     */
+    final public function begin(): string
+    {
+        $this->beginExecuted = true;
+
+        $renderBegin = $this->runBegin();
+        $stack = self::getContextStack();
+
+        $stack[] = $this;
+
+        self::$stack?->offsetSet(self::getContextId(), $stack);
+
+        return $renderBegin;
+    }
+
+    /**
+     * Ends the most recent begin/end rendering block for the current class.
+     *
+     * @throws LogicException if no matching `begin()` call is found.
+     * @throws RuntimeException if the tag class does not match the expected type.
+     *
+     * @return string Rendered HTML tag string.
+     */
+    final public static function end(): string
+    {
+        $key = self::getContextId();
+        $stack = self::getContextStack();
+
+        if ($stack === []) {
+            throw new LogicException(
+                Message::UNEXPECTED_END_CALL_NO_BEGIN->getMessage(static::class),
+            );
+        }
+
+        $tag = array_pop($stack);
+
+        if ($stack === []) {
+            self::$stack?->offsetUnset($key);
+        } else {
+            self::$stack?->offsetSet($key, $stack);
+        }
+
+        $tagClass = $tag::class;
+
+        if ($tagClass !== static::class) {
+            throw new RuntimeException(
+                Message::TAG_CLASS_MISMATCH_ON_END->getMessage($tagClass, static::class),
+            );
+        }
+
+        return $tag->render();
+    }
 
     /**
      * Cleans up the output after rendering the block element.
@@ -92,6 +176,14 @@ abstract class BaseBlock extends BaseTag
     protected function afterRun(string $result): string
     {
         return LineBreakNormalizer::normalize(parent::afterRun($result));
+    }
+
+    /**
+     * Indicates whether `begin()` was executed for this instance.
+     */
+    protected function isBeginExecuted(): bool
+    {
+        return $this->beginExecuted;
     }
 
     /**
@@ -121,5 +213,33 @@ abstract class BaseBlock extends BaseTag
     protected function runBegin(): string
     {
         return Html::begin($this->getTag(), $this->getAttributes());
+    }
+
+    /**
+     * Returns the identifier for the current execution context.
+     *
+     * @phpstan-return Fiber<mixed, mixed, mixed, mixed>|stdClass
+     */
+    private static function getContextId(): Fiber|stdClass
+    {
+        return Fiber::getCurrent() ?? self::$mainThread ??= new stdClass();
+    }
+
+    /**
+     * Returns the begin/end stack for the current execution context.
+     *
+     * @phpstan-return array<array-key, static>
+     */
+    private static function getContextStack(): array
+    {
+        self::$stack ??= new WeakMap();
+
+        $key = self::getContextId();
+
+        if (self::$stack->offsetExists($key) === false) {
+            self::$stack->offsetSet($key, []);
+        }
+
+        return self::$stack->offsetGet($key);
     }
 }
