+++ src/Element/BaseVoid.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace UIAwesome\Html\Core\Element;
 
+use BackedEnum;
 use UIAwesome\Html\Attribute\Global\{
     CanBeHidden,
     HasAccesskey,
@@ -19,15 +20,16 @@ use UIAwesome\Html\Attribute\Global\{
     HasTitle,
     HasTranslate,
 };
+use UIAwesome\Html\Contracts\Attribute\AttributesInterface;
+use UIAwesome\Html\Contracts\Element\VoidInterface;
 use UIAwesome\Html\Core\Base\BaseTag;
 use UIAwesome\Html\Core\Html;
-use UIAwesome\Html\Interop\VoidInterface;
 use UIAwesome\Html\Mixin\HasAttributes;
 
 /**
  * Provides the base implementation for void HTML elements.
  *
- * Subclasses return a {@see VoidInterface} tag and inherit attribute handling for rendering through
+ * Subclasses return a {@see BackedEnum} tag and inherit attribute handling for rendering through
  * {@see Html::void()}.
  *
  * @link https://developer.mozilla.org/en-US/docs/Glossary/Void_element
@@ -35,7 +37,7 @@ use UIAwesome\Html\Mixin\HasAttributes;
  * @copyright Copyright (C) 2025 Terabytesoftw.
  * @license https://opensource.org/license/bsd-3-clause BSD 3-Clause License.
  */
-abstract class BaseVoid extends BaseTag
+abstract class BaseVoid extends BaseTag implements AttributesInterface, VoidInterface
 {
     use CanBeHidden;
     use HasAccesskey;
@@ -55,17 +57,17 @@ abstract class BaseVoid extends BaseTag
     /**
      * Returns the tag instance representing the void element.
      *
-     * @return VoidInterface Tag instance for the void element.
+     * @return BackedEnum Tag instance for the void element.
      *
      * Usage example:
      * ```php
-     * public function getTag(): VoidInterface
+     * public function getTag(): BackedEnum
      * {
-     *     return \UIAwesome\Html\Interop\Void::IMG;
+     *     return \UIAwesome\Html\Interop\Voids::IMG;
      * }
      * ```
      */
-    abstract protected function getTag(): VoidInterface;
+    abstract protected function getTag(): BackedEnum;
 
     /**
      * Renders the void element.
